# Performance improvements made in response to profiling