--- a/Tests/Benchmarks/NestedClass.cs
+++ b/Tests/Benchmarks/NestedClass.cs
@@ -6,8 +6,15 @@
 
         public Nested B { get; set; }
 
-        public static Nested Create(int i)
+        public Nested C { get; set; }
+
+        public static Nested Create(int i, ref int objectCount)
         {
+            if (i < 0)
+            {
+                return null;
+            }
+
             if (i == 0)
             {
                 return new Nested();
@@ -15,8 +22,9 @@
 
             return new Nested
             {
-                A = i,
-                B = Create(i - 1)
+                A = ++objectCount,
+                B = Create(i - 1, ref objectCount),
+                C = Create(i - 2, ref objectCount),
             };
         }
     }