… Tests/Approval.Tests/ApprovedApi/FluentAssertions/netstandard2.0.approved.txt
… Tests/Approval.Tests/ApprovedApi/FluentAssertions/netstandard2.1.approved.txt
… Tests/FluentAssertions.Specs/Common/TypeExtensions.cs