This PR makes some performance improvements based on profiling with #1419's changes, testing DataTable objects with many records:

- Updated TypeExtensions.cs to short-circuit GetClosedGenericInterfaces if the type isn't an object type.
- Updated AssertionScope.cs to support lazily resolving the context string, storing a Lazy<string> promise instead for as long as needed. Added internal method ResolveContext() to be used by AssertionRuleEquivalencyStep.cs before recursing, so that the rule doesn't get picked up as the context. Updated EquivalencyValidator.cs to set a promise rather than actually computing the Context value on-the-spot every time a new scope is created.
- Updated ContextDataItems.cs to replace DataItems inline rather than removing/re-adding list elements. This means that the list won't always have the most recently added item at the back, but that doesn't seem to be important. Didn't break any tests, anyway.
- Updated CyclicReferenceDetector.cs to use a Dictionary keyed on the target object instead of a List for its object references, so that when the number of objects gets extremely large, performance doesn't suffer due to linear scans. Added an internal Object accessor to ObjectReference.cs.

## IMPORTANT 

* [x] The code complies with the [Coding Guidelines for C#](https://www.csharpcodingguidelines.com/).
* [x] The changes are covered by a new or existing set of unit tests which follow the Arrange-Act-Assert syntax such as is used [in this example](https://github.com/fluentassertions/fluentassertions/blob/daaf35b9b59b622c96d0c034e8972a020b2bee55/Tests/FluentAssertions.Shared.Specs/BasicEquivalencySpecs.cs#L33).
* [x] ~If the contribution adds a feature or fixes a bug, please update [**the release notes**](https://github.com/fluentassertions/fluentassertions/tree/master/docs), which are published on the [website](https://fluentassertions.com/releases).~
* [x] ~If the contribution changes the public API the changes needs to be included by running `AcceptApiChanges.ps1`/`AcceptApiChanges.sh`.~
* [x] ~If the contribution affects [the documentation](https://github.com/fluentassertions/fluentassertions/tree/develop/docs/_pages), please include your changes in this pull request so the documentation will appear on the [website](https://www.fluentassertions.com/introduction).~