@lg2de 2020-11-17
Would be great to have some measurement on the improvements, e.g. using BenchmarkDotNet with example data before and after these changes.