@logiclrd 2020-11-17
That's reasonable. These changes were originally a part of #1419, in which I ran into severe performance issues when trying to work with very large DataTable objects. The two main performance killers were that:

1. Even though it was never meant to, the code is rediscovering the source line of the context every time the context changes, _not_ only when it's about to report a failure. This meant that if the object graph had a million items in it, it was parsing the stack trace and then scanning the source code to find the expression at the line/column in the first non-FluentAssertions frame over and over, exactly the same result each time, a million times.

2. The cyclic reference checking code uses a linear scan for finding matching objects. This causes an O(n^2) performance characteristic, which becomes noticeable with very large object graphs.

I can't remember exactly what DataTable size I did the calculations with, I believe it was 100,000 rows, but with the changes here I observed a 60-fold increase in performance and also confirmed a performance characteristic better than O(n^2).