@logiclrd 2020-11-17
I have added a performance benchmark that sets up large objects graphs with a range of sizes and passes them to `.BeEquivalentTo`. I am running it now on the pre-improvements code. I'm not sure how far it'll get, though, because at a graph size of 2,583 objects, it's taking approximately 1.8 _seconds_ per `BeEquivalentTo` call. This benchmark runs up to object graph sizes exceeding 3 million objects. :-P

I'll let it run for a bit and then show what results it has managed to achieve, then I'll run it with the performance changes.