@logiclrd 2020-11-17
Old code:
```
N = 16 ; Graph size: 2583 objects

Mean = 1.807 s, StdErr = 0.009 s (0.50%), N = 59, StdDev = 0.070 s
Min = 1.707 s, Q1 = 1.761 s, Median = 1.788 s, Q3 = 1.820 s, Max = 2.059 s
IQR = 0.059 s, LowerFence = 1.672 s, UpperFence = 1.909 s
ConfidenceInterval = [1.775 s; 1.838 s] (CI 99.9%), Margin = 0.031 s (1.74% of Mean)
Skewness = 1.6, Kurtosis = 5.58, MValue = 2

N = 18 ; Graph size: 6764 objects

Mean = 5.152 s, StdErr = 0.026 s (0.51%), N = 19, StdDev = 0.114 s
Min = 4.985 s, Q1 = 5.062 s, Median = 5.140 s, Q3 = 5.220 s, Max = 5.364 s
IQR = 0.158 s, LowerFence = 4.825 s, UpperFence = 5.457 s
ConfidenceInterval = [5.050 s; 5.254 s] (CI 99.9%), Margin = 0.102 s (1.99% of Mean)
Skewness = 0.42, Kurtosis = 1.97, MValue = 2

N = 20 ; Graph size: 17710 objects

WorkloadWarmup   1: 1 op, 16910015000.00 ns, 16.9100 s/op
WorkloadWarmup   2: 1 op, 16080022500.00 ns, 16.0800 s/op
WorkloadWarmup   3: 1 op, 16481630700.00 ns, 16.4816 s/op
WorkloadWarmup   4: 1 op, 16336174700.00 ns, 16.3362 s/op
WorkloadWarmup   5: 1 op, 15959460200.00 ns, 15.9595 s/op
...
```

At these scales, the O(n^2) characteristic of the cyclic reference check isn't terribly obvious because the results are utterly dominated by the repeated computation of the subject (with FluentAssertions does by locating & inspecting the caller's source code). But, when I crunch the numbers further down, it can be seen to be growing noticeably faster than the number of objects.

New code:
```
|         Method |  N |         Mean |      Error |     StdDev |        Gen 0 |       Gen 1 |     Gen 2 |   Allocated |
|--------------- |--- |-------------:|-----------:|-----------:|-------------:|------------:|----------:|------------:|
| BeEquivalentTo | 16 |     52.18 ms |   1.018 ms |   0.952 ms |    8000.0000 |   2000.0000 |  571.4286 |    48.59 MB |
| BeEquivalentTo | 18 |    139.02 ms |   2.726 ms |   3.348 ms |   21250.0000 |   5250.0000 | 1000.0000 |   128.38 MB |
| BeEquivalentTo | 20 |    370.64 ms |   6.778 ms |   6.008 ms |   57000.0000 |  15000.0000 | 1000.0000 |   339.28 MB |
| BeEquivalentTo | 24 |  2,603.46 ms |  50.658 ms |  62.212 ms |  394000.0000 | 100000.0000 | 1000.0000 |  2379.35 MB |
| BeEquivalentTo | 28 | 22,344.94 ms | 442.011 ms | 590.072 ms | 2757000.0000 | 691000.0000 | 1000.0000 | 16608.43 MB |
```

Test sizes:

N = 16 ; Graph size: 2583 objects
N = 18 ; Graph size: 6764 objects
N = 20 ; Graph size: 17710 objects
N = 24 ; Graph size: 121392 objects
N = 28 ; Graph size: 832039 objects

With the new code, the performance proportional to the number of objects is:
```
|  N | Objects | Time/Object |
|----|---------|-------------|
| 16 |    2583 |     20.2 us |
| 18 |    6764 |     20.5 us |
| 20 |   17710 |     20.9 us |
| 24 |  121392 |     21.4 us |
| 28 |  832039 |     26.9 us |
```

The average time/object for the old code, for the duration of testing I let it complete, was:
```
|  N | Objects | Time/Object | Ratio |
|----|---------|-------------|-------|
| 16 |    2583 |    699.6 us |  35x  |
| 18 |    6764 |    761.7 us |  37x  |
| 20 |   17710 |    923.4 us |  44x  |
```