@logiclrd 2020-11-19
This latest commit seems to work. The trickiest part was the `async` machinery, which gums up the works with stack frames that aren't in `FluentAssertions` but do need to be ignored. Because of the fact that `action.Should().Throw(..)` type assertions have this exact same structure, where the stack frames at the bottom of the stack will find `action.Should()` but the message should be finding the subject within the "..", I had to create a mechanism for assertions on delegates to indicate that the CallerIdentifier logic should find the _second_ instance of user code calling into FluentAssertions and not the first. This only activates if the `action.Should()` is the very first call into `FluentAssertions` on the call stack.