@dennisdoomen 2020-11-22
> I am getting unit test failures on `develop` currently. The same tests are failing after my changes here, but hopefully the main line will see fixes onto which this branch can be rebased.

I've just tried on AppVeyor and locally, but both work without problems. Are you sure? Which one are failing?